Move the DoubleSlashChecker to a separate file as a part of custom module
