+++ Rakefile
@@ -14,6 +14,7 @@ require 'colorator'
 # Load ruby files with helper methods from the 'rakelib/' directory
 require_relative 'rakelib/link-checker.rb'
 require_relative 'rakelib/converter.rb'
+require_relative 'rakelib/double-slash-check.rb'
 
 desc "Same as 'rake', 'rake preview'"
 task default: %w[preview]
