+++ rakelib/link-checker.rb
@@ -1,26 +1,5 @@
 # frozen_string_literal: true
 
-# Custom class to find double forward slashes in URLs.
-
-class DoubleSlashCheck < ::HTMLProofer::Check
-
-  def slash?
-    return false if @link.href.nil?
-    @link.href.match /\w\/\//
-  end
-
-  def run
-    @html.css('a').each do |node|
-      @link = create_element(node)
-      line = node.line
-
-      if slash?
-        return add_issue("Remove double forward slashes from URLs", line: line)
-      end
-    end
-  end
-end
-
 # Helper methods to configure  and run html-proofer
 module LinkChecker
   # Run html-proofer to check the generated HTML pages
